+++ src/statistics.h
@@ -8,7 +8,7 @@ extern "C" {
 #include "timeframe.h"
 #include <fftw3.h>
 
-double mean(int length, const double *array);
+double mean(size_t length, const double *array);
 
 void susceptibility_lin(const double complex *isf, double *spike_train,
                         const TimeFrame *time_frame, double complex *suscept,
